crearTorneo

+++ src/basedatos/dao/PartidaDAO.java
@@ -31,7 +31,7 @@ public class PartidaDAO {
                         sd.format(p.getTimeInicio()) + "'," + p.isPublica() + ")";
             }
             else{
-                if (p.getTorneoId() == 0) {
+                if (p.getTorneoId() == BigInteger.valueOf(0)) {
                     insertPartida = "INSERT INTO partida (publica) VALUES (" + (p.isPublica() ? 1 : 0) + ")";
                 } else {
                     insertPartida = "INSERT INTO partida (fase_num, fase_torneo,publica) VALUES (" + p.getFaseNum() + "," + p.getTorneoId() +",1)";
